Feature/exception

+++ src/main/java/com/example/service/implementation/VendorServiceImpl.java
@@ -1,11 +1,13 @@
 package com.example.service.implementation;
 
 import com.example.entity.Vendor;
+import com.example.exception.VendorNotFoundException;
 import com.example.repository.VendorRepository;
 import com.example.service.VendorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class VendorServiceImpl implements VendorService {
@@ -24,14 +26,28 @@ public class VendorServiceImpl implements VendorService {
     }
 
     @Override
-    public void updateVendorName(Vendor vendor) {
+    public void updateVendorName(int id, String name) {
 
-        vendorRepository.save(vendor);
+        Optional<Vendor> vendorOptional = vendorRepository.findById(id);
+        if(vendorOptional.isPresent()){
+            Vendor vendorForDB = new Vendor();
+            vendorForDB.setName(name);
+            vendorRepository.save(vendorForDB);
+        }
+        else{
+            throw new VendorNotFoundException(id);
+        }
     }
 
     @Override
     // TODO : Handle key constraint violation exception
     public void deleteVendor(int id) {
-        vendorRepository.deleteById(id);
+        Optional<Vendor> vendorOptional = vendorRepository.findById(id);
+        if(vendorOptional.isPresent()){
+            vendorRepository.deleteById(id);
+        }
+        else{
+            throw new VendorNotFoundException(id);
+        }
     }
 }
